refactor: start to export interfaces
