+++ src/web/src/views/workspace/commandContent/index.ts
@@ -1,2 +1,2 @@
 export { default as WSEditorCommandContent, DecodeResponseCommand } from "./WSEditorCommandContent";
-export type { Command, Resource, ResponseCommand, ResponseCommands, Example, Plane } from "./WSEditorCommandContent";
+export type { Command, Resource, ResponseCommand, ResponseCommands, Example, Plane } from "../interfaces";
