+++ src/web/src/views/workspace/interfaces/index.ts
@@ -1,0 +1,3 @@
+export type { Plane, Example, Resource, Command, ResponseCommand, ResponseCommands } from "./commands";
+
+export type { CommandGroup, ResponseCommandGroup, ResponseCommandGroups } from "./commandGroups";
